+++ CHANGELOG.md
@@ -25,6 +25,11 @@ Changes are kept under subsections:
 Only major refactoring improvements or functional changes should be noted.
 
 --------------------------------------------------------------------------------
+## [3.2.0] - 2025-07-09 - Upload subsets as zip and unzip
+### New
+- Allow for istex subset zip upload and extract
+- Show subset papers on web list
+
 ## [3.1.0] - 2025-06-18 - Update stats
 ### Changed
 - Web button for stats update